Add category element to form section
