--- src/components/selectCategories.js
@@ -1,85 +1,0 @@
-import * as React from 'react';
-import { useTheme } from '@mui/material/styles';
-import OutlinedInput from '@mui/material/OutlinedInput';
-import InputLabel from '@mui/material/InputLabel';
-import MenuItem from '@mui/material/MenuItem';
-import FormControl from '@mui/material/FormControl';
-import Select from '@mui/material/Select';
-
-const ITEM_HEIGHT = 48;
-const ITEM_PADDING_TOP = 8;
-const MenuProps = {
-  PaperProps: {
-    style: {
-      maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
-      width: 250,
-    },
-  },
-};
-
-const categories = [
-  'Action',
-  'Adventure',
-  'Fantasy',
-  'Horror',
-  'Classic',
-  'History',
-  'Detective and Mystery',
-  'Romance',
-  'Sci-Fi',
-  'Humor',
-  'Tragedy',
-  'Computer Programming',
-  'Documentary',
-];
-
-function getStyles(category, categoryName, theme) {
-  return {
-    fontWeight:
-      categoryName.indexOf(category) === -1
-        ? theme.typography.fontWeightRegular
-        : theme.typography.fontWeightMedium,
-  };
-}
-
-export default function MultipleSelect() {
-  const theme = useTheme();
-  const [categoryName, setCategoryName] = React.useState([]);
-
-  const handleChange = (event) => {
-    const {
-      target: { value },
-    } = event;
-    setCategoryName(
-      // On autofill we get a stringified value.
-      typeof value === 'string' ? value.split(',') : value,
-    );
-  };
-
-  return (
-    <div>
-      <FormControl sx={{ m: 1, width: 300 }}>
-        <InputLabel id="demo-multiple-name-label">Category</InputLabel>
-        <Select
-          labelId="demo-multiple-name-label"
-          id="demo-multiple-name"
-          multiple
-          value={categoryName}
-          onChange={handleChange}
-          input={<OutlinedInput label="Category" />}
-          MenuProps={MenuProps}
-        >
-          {categories.map((category) => (
-            <MenuItem
-              key={category}
-              value={category}
-              style={getStyles(category, categoryName, theme)}
-            >
-              {category}
-            </MenuItem>
-          ))}
-        </Select>
-      </FormControl>
-    </div>
-  );
-}
